Create Script to Terminate the user and email files
Pull Data from SQL tables
Check the status and disable it on a valid date. Treat exceptions.

Remove groups, manager name, update desc, notes, diable the account, move the OU

Update in an excel file, create a folder for each user,

Send email to notify removal of groups and also to cancel subscriptions.

Send an email to the group 

ALl functions in one.